ingestion/requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
pypdf>=3.0.0
google-genai>=1.0.0
